Moves guitars data management to Repository class.
