+++ MusicStore/MusicStore/Controllers/GuitarController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using MusicStore.Data;
 using MusicStore.Models;
 using System;
 using System.Collections.Generic;
@@ -11,8 +12,9 @@ namespace MusicStore.Controllers
     public class GuitarController : Controller
     {
         public ActionResult Index()
-        {            
-            return View(GuitarColletction);
+        {
+            var guitars = Repository.GuitarColletction;
+            return View(guitars);
         }
 
         public ActionResult Details(int id)
@@ -75,33 +77,6 @@ namespace MusicStore.Controllers
             {
                 return View();
             }
-        }
-
-
-        private static List<Guitar> guitars = null;
-        private static List<Guitar> GuitarColletction
-        {
-            get
-            {
-                if (guitars == null)
-                {
-                    guitars = new List<Guitar>();
-                    guitars.Add(new Guitar() { Name = "Standard Strat.", ImagePath = "img/img-1.jpg", Price = 1600 });
-                    guitars.Add(new Guitar() { Name = "Rosewood Std.", ImagePath = "img/img-2.jpg", Price = 1850 });
-                    guitars.Add(new Guitar() { Name = "Limited Edition", ImagePath = "img/img-3.jpg", Price = 2665 });
-                    guitars.Add(new Guitar() { Name = "Standard Mapple", ImagePath = "img/img-4.jpg", Price = 1500 });
-                }
-
-                return guitars;
-            }
-        }
-
-        private IEnumerable<Guitar> GetGuitarsByName(string filter)
-        {
-            var guitars = from guitar in GuitarColletction
-                          where guitar.Name.Contains(filter.Trim())
-                          select guitar;
-            return guitars.ToList();
         }
     }
 }
